Read coverage counts as Int64, narrowing only at the vendored boundary
Every `x86` leg that runs with coverage errors on whatever test item happens to
be running when the counters are collected:

    OverflowError: overflow parsing "2345022144"
      at readfile (packages/CoverageTools/src/lcov.jl:99)

`collect_coverage_data!` writes the runtime's counters with
`jl_write_coverage_data` and reads them straight back with
`CoverageTools.LCOV.readfile`. Julia's counters are 64 bit whatever the word
size, and a line in a hot loop passes `typemax(Int32)` easily, but the count was
parsed as `Int` — `Int32` on a 32 bit platform — so the value the runtime had
just written did not fit.

The fix cannot widen `CovCount`: `packages/` holds git subtrees that are never
edited by hand, and `scripts/update_vendored_packages.jl` deliberately skips
CoverageTools, so there is no re-vendoring route either. So the file is read
here instead. `shared/coverage_counts.jl` parses the same `SF:`/`DA:` lines that
`LCOV.readfile` does, accumulating in `Int64`, and narrows a count only where it
crosses into a vendored `FileCoverage` — where it saturates rather than throwing,
because knowing a line ran at least 2147483647 times beats losing the run over
the exact figure. The controller applies the same narrowing to counts arriving
over the wire, so a 64 bit test process reporting to a 32 bit controller cannot
throw an `InexactError` either.

The protocol's `FileCoverage` now carries `Int64`, so the JSON payload does not
depend on either side's word size. `TestrunResultFileCoverage` stays at `Int`:
what reaches it comes out of the vendored `merge_coverage_counts`, which produces
`Int` whatever we declare.

+++ shared/testserver_protocol.jl
@@ -62,15 +62,17 @@ TestMessage(message, location) = TestMessage(message, missing, missing, location
     timeoutMs::Union{Missing,Float64}
 end
 
+# `Int64` rather than `Int`: a controller and a test process can run at different word
+# sizes, and Julia's coverage counters are 64 bit on both.
 struct FileCoverage <: JSONRPC.Outbound
     uri::String
-    coverage::Vector{Union{Int,Nothing}}
+    coverage::Vector{Union{Int64,Nothing}}
 end
 
 function FileCoverage(d::Dict)
     return FileCoverage(
         d["uri"],
-        Union{Int,Nothing}[i for i in d["coverage"]]
+        Union{Int64,Nothing}[i for i in d["coverage"]]
     )
 end
